initial stab at templatizing gbrowse.  templates live in conf/templates/$appname (e.g. conf/templates/gbrowse/).
i've converted over the navigation table, the scroll/zoom tool, and the
search box to templates to give people an idea how it works.

the next logical step (aside from templatizing all other html generating
code), is to templatize the lang .pm files into templates that can just be
called in with an [% INCLUDE "mylang.tt2" %] directive in the toplevel
template.  this will simplify the logic code by reducing the number of calls
that need to be made to $CONF just to ferry the translations into the
template.

i've also started adding better debug log support that is written to
$CONF_DIR/gbrowse.log.  see conf/log4perl.conf for details.

this introduces prereqs on the Template toolkit and Log::Log4perl.  you can
get them e.g.

  cpan> install Template Log::Log4perl