Fixed an issue when appending tracks to the panel in the add_tracks method.  I was using:
          $(panel_id).innerHTML += html;

But that was causing all of the previous tracks javascript to be read (without executing register_track method).  So the old tracks would lose their ability to get new images.

The change is that now, the html is read into a specially created tmp_element which is then appended to the panel.  The new div element is then moved from the tmp_element to the panel element and the tmp_element is removed (which does not remove it from memory but at least it cleans up the dom object).  This preserves the other divs.