Fixed a problem where the imported (remote and uploaded) annotations were having odd URLs created for the feature links.  The desired behavior was for the imported features not to have links at all.
The problem was caused by the call to $panel->add_track() which isn't needed for a feature_file (aka uploaded/remote annotations).  Essentially, a second and useless "track" was being created on the panel.  This incorrect "track" was being used to determine if the track was a feature file and since it never could be, that check was failing.